Indices API Reference

ElasticsearchClient.API.Indices.create — Function

Creates an index with optional settings and mappings.

index::String: The name of the index

wait_for_active_shards::String: Set the number of active shards to wait for before the operation returns.

timeout::DateTime: Explicit operation timeout

master_timeout::DateTime: Specify timeout for connection to master

headers::Dict: Custom HTTP headers

body::Union{NamedTuple,Dict}: The configuration for the index (settings and mappings)

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-create-index.html

ElasticsearchClient.API.Indices.delete — Function

Deletes an index.

index::Union{String,Vector{String}}: A comma-separated list of indices to delete; use _all or * string to delete all indices

timeout::DateTime: Explicit operation timeout

master_timeout::DateTime: Specify timeout for connection to master

ignore_unavailable::Bool: Ignore unavailable indexes (default: false)

allow_no_indices::Bool: Ignore if a wildcard expression resolves to no concrete indices (default: false)

expand_wildcards::String: Whether wildcard expressions should get expanded to open, closed, or hidden indices (options: open, closed, hidden, none, all)

headers::Dict: Custom HTTP headers

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-delete-index.html

ElasticsearchClient.API.Indices.exists_alias — Function

Returns information about whether a particular alias exists.

name::Union{String,Vector{String}}: A comma-separated list of alias names to return

index::Union{String,Vector{String}}: A comma-separated list of index names to filter aliases

ignore_unavailable::Bool: Whether specified concrete indices should be ignored when unavailable (missing or closed)

allow_no_indices::Bool: Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)

expand_wildcards::String: Whether to expand wildcard expression to concrete indices that are open, closed or both. (options: open, closed, hidden, none, all)

local::Bool: Return local information, do not retrieve the state from master node (default: false)

headers::Dict: Custom HTTP headers

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-aliases.html

ElasticsearchClient.API.Indices.put_alias — Function

Creates or updates an alias.

index::Union{String,Vector{String}}: A comma-separated list of index names the alias should point to (supports wildcards); use _all to perform the operation on all indices.

name::String: The name of the alias to be created or updated

timeout::DateTime: Explicit timestamp for the document

master_timeout::DateTime: Specify timeout for connection to master

headers::Dict: Custom HTTP headers

body::Union{NamedTuple,Dict}: The settings for the alias, such as routing or filter

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-aliases.html

ElasticsearchClient.API.Indices.get — Function

Returns information about one or more indices.

index::Union(String,Vector{String}}: A comma-separated list of index names

local::Bool: Return local information, do not retrieve the state from master node (default: false)

ignore_unavailable::Bool: Ignore unavailable indexes (default: false)

allow_no_indices::Bool: Ignore if a wildcard expression resolves to no concrete indices (default: false)

expand_wildcards::String: Whether wildcard expressions should get expanded to open or closed indices (default: open) (options: open, closed, hidden, none, all)

features::String: Return only information on specified index features (options: aliases, mappings, settings)

flat_settings::Bool: Return settings in flat format (default: false)

include_defaults::Bool: Whether to return all default setting for each of the indices.

master_timeout::DateTime: Specify timeout for connection to master

headers::Dict: Custom HTTP headers

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-get-index.html

ElasticsearchClient.API.Indices.exists — Function

Returns information about whether a particular index exists.

index::Union{String,Vector{String}}: A comma-separated list of index names

local::Bool: Return local information, do not retrieve the state from master node (default: false)

ignore_unavailable::Bool: Ignore unavailable indexes (default: false)

allow_no_indices::Bool: Ignore if a wildcard expression resolves to no concrete indices (default: false)

expand_wildcards::String: Whether wildcard expressions should get expanded to open or closed indices (default: open) (options: open, closed, hidden, none, all)

flat_settings::Bool: Return settings in flat format (default: false)

include_defaults::Bool: Whether to return all default setting for each of the indices.

headers::Dict: Custom HTTP headers

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-exists.html

ElasticsearchClient.API.Indices.refresh — Function

Performs the refresh operation in one or more indices.

index::Union{String,Vector{String}}: A comma-separated list of index names; use _all or empty string to perform the operation on all indices

ignore_unavailable::Bool: Whether specified concrete indices should be ignored when unavailable (missing or closed)

allow_no_indices::Bool: Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)

expand_wildcards::String: Whether to expand wildcard expression to concrete indices that are open, closed or both. (options: open, closed, hidden, none, all)

headers::Dict: Custom HTTP headers

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-refresh.html

ElasticsearchClient.API.Indices.update_aliases — Function

Updates index aliases.

timeout::DateTime: Request timeout

master_timeout::DateTime: Specify timeout for connection to master

headers::Dict: Custom HTTP headers

body::Union{NamedTuple,Dict}: The definition of actions to perform (Required)

See https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-aliases.html